compiler: fix comp. of func calls for func values of a defined type When compiling a piece of code where a function value is called, the compiler panics if the function value's type is a defined type, and not just a type literal (function signature): The type assertion (*types.Signature) fails, because the type of the func value is a *types.Named.
This patch fixes this by using the type's underlying type, so that a
types.Named is properly turned into its underlying types.Signature,
before the type assertion takes place.
It takes advantage of the property that all types have an underlying type
(both are the same, if a type is not named).

Fixes #320