interp: improve soundness of loads and stores
Previously, whether a value could be safely loaded from or stored to was
based on some heuristics. This commit fixes that by making sure the
underlying global is really not part of the dirty set (which must be
operated on at runtime instead of compile time).

Together with the previous commit, this commit gets a number of standard
library packages to compile:

  * crypto/dsa
  * crypto/ed25519
  * crypto/elliptic
  * crypto/rand
  * go/constant
  * math/big