runtime (gc): bundle small allocations together
This reduces the overhead of small allocations.
Previously a pointer sized allocation would take up an entire block - leaving 3 pointer-widths of wasted space.
Small allocations are now bundled together, and are treated as a single object for garbage-collection purposes.
These could be tracked seperately, but not without paying for space to store additional garbage collector metadata.