Programming model updated to use machine generated descriptions.
This commit changes several files related to the raspberry pi.
It assumes that if you want to change the configuration of the
linkerscript or other parameters, you will override the one
provided with your own local .json file. (Thanks ayke.)

There is a sample program to use with these changes that can
added to tinygo-zoo once these changes are merged and
deployed.